Move scheduling functions to their own module
Which gets called by cli.py
